+++ lib/clients/metaApi/clientApi_client.py
@@ -5,6 +5,7 @@ from typing_extensions import TypedDict
 from typing import List
 from datetime import datetime
 import asyncio
+from ...logger import LoggerManager
 
 
 class TypeHashingIgnoredFieldLists(TypedDict):
@@ -37,11 +38,13 @@ class ClientApiClient(MetaApiClient):
         """
         super().__init__(http_client, domain_client)
         self._host = f'https://mt-client-api-v1'
+        self._retryIntervalInSeconds = 1
         self._ignoredFieldListsCache = {
             'lastUpdated': 0,
             'data': None,
             'requestPromise': None
         }
+        self._logger = LoggerManager.get_logger('ClientApiClient')
 
     async def get_hashing_ignored_field_lists(self, region: str) -> HashingIgnoredFieldLists:
         """Retrieves hashing ignored field lists.
@@ -59,22 +62,26 @@ class ClientApiClient(MetaApiClient):
             else:
                 future = asyncio.Future()
                 self._ignoredFieldListsCache['requestPromise'] = future
-                host = await self._domainClient.get_url(self._host, region)
-                opts = {
-                    'url': f'{host}/hashing-ignored-field-lists',
-                    'method': 'GET',
-                    'headers': {
-                        'auth-token': self._token
-                    }
-                }
-
-                try:
-                    response = await self._httpClient.request(opts, 'get_hashing_ignored_field_lists')
-                    self._ignoredFieldListsCache = {
-                        'lastUpdated': datetime.now().timestamp(), 'data': response, 'requestPromise': None}
-                    future.set_result(response)
-                except Exception as err:
-                    future.set_exception(err)
-                    raise err
+                is_cache_updated = False
+                while not is_cache_updated:
+                    try:
+                        host = await self._domainClient.get_url(self._host, region)
+                        opts = {
+                            'url': f'{host}/hashing-ignored-field-lists',
+                            'method': 'GET',
+                            'headers': {
+                                'auth-token': self._token
+                            }
+                        }
+                        response = await self._httpClient.request(opts, 'get_hashing_ignored_field_lists')
+                        self._ignoredFieldListsCache = {
+                            'lastUpdated': datetime.now().timestamp(), 'data': response, 'requestPromise': None}
+                        future.set_result(response)
+                        is_cache_updated = True
+                        self._retryIntervalInSeconds = 1
+                    except Exception as err:
+                        self._logger.error('Failed to update hashing ignored field list', err)
+                        self._retryIntervalInSeconds = min(self._retryIntervalInSeconds * 2, 300)
+                        await asyncio.sleep(self._retryIntervalInSeconds)
 
         return self._ignoredFieldListsCache['data']